+++ src/satellite/orbcomm/orbcomm.service.ts
@@ -39,6 +39,7 @@ export class OrbcommService {
   @Cron(CronExpression.EVERY_10_SECONDS)
   async checkMessages() {
     console.log('UPDATE MESSAGES PROCESS...');
+
     try {
       findMessagesByOrbcommStatus(this.prisma)
         .then(createListOfFwdIds)
@@ -51,5 +52,9 @@ export class OrbcommService {
     } catch (error) {
       return error;
     }
+  }
+
+  async downloadMessages() {
+    console.log('DOWNLOAD MESSAGES PROCESS....');
   }
 }
